feat: Refactor i18n imports and update routing paths across components
