+++ src/app/(routes)/auditoria/view/page.js
@@ -1,5 +1,5 @@
 "use client";
-import "@/i18n";
+import "@i18n/i18n";
 import { useRouter } from "next/navigation";
 import VisualizarAuditoria from '@/app/components/Auditoria/visualizarAuditoria';
 import Alert from '@components/common/alerts/Alert';
