+++ src/app/(routes)/aula/new/page.js
@@ -1,7 +1,7 @@
 "use client";
 
 import CrearForm from '@components/ClassroomAssignment/CrearForm';
-import "@/i18n";
+import "@i18n/i18n";
 import Alert from '@components/common/alerts/Alert';
 import PropTypes from 'prop-types';
