+++ src/app/Main.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { BrowserRouter } from 'react-router-dom';
 import App from './App.jsx';
-import './i18n';
+import './i18n/i18n.js';
 import './index.css';
 
 ReactDOM.createRoot(document.getElementById('root')).render(
