+++ src/app/components/Recursos/CrearRecursos.jsx
@@ -51,9 +51,9 @@ const CrearRecursos = ({ alert = null, onSubmit = () => {} }) => {
       <div className="flex items-center mb-6 px-12 py-8">
         <ChevronLeft
         className="w-6 h-6 text-black_text mr-2 cursor-pointer"
-        onClick={() => router.push('/recursos/visualizar')}
+        onClick={() => router.push('/recursos/view')}
         />
-        <h2 className="text-2xl font-bold text-gray-800">Registrar Material TIC</h2>
+        <h2 className="text-2xl font-bold text-gray-800">Crear Material TIC</h2>
       </div>
 
       {/* Formulario */}
