+++ src/app/layout.js
@@ -1,5 +1,5 @@
 import './globals.css';
-import './i18n'
+import './i18n/i18n'
 
 export const metadata = {
   title: 'Next.js',
